+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
 All notable changes to the "vscode-glsl-linter" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## 1.1.1
+- This is not the initial release, but I noticed the changelog file right now
 
-## [Unreleased]
-- Initial release
+## 1.2.0
+- Fixed non-empty `glsl-linter.validatorArgs` string causing the validator to fail silently
+- `glsl-linter.validatorArgs` is now an array
+- Error handling improvements